+++ setup.cfg
@@ -16,7 +16,12 @@ install_requires =
     spacy>=3.2.0,<3.3.0
 
 [options.entry_points]
-spacy_factories = 
+spacy_architectures =
+    spacy-experimental.Bilinear.v1 = spacy_experimental.biaffine_parser.bilinear:build_bilinear
+    spacy-experimental.PairwiseBilinear.v1 = spacy_experimental.biaffine_parser.pairwise_bilinear:build_pairwise_bilinear
+spacy_factories =
+    experimental_arc_labeler = spacy_experimental.biaffine_parser.arc_labeler:make_arc_labeler
+    experimental_arc_predicter = spacy_experimental.biaffine_parser.arc_predicter:make_arc_predicter
     experimental_edit_tree_lemmatizer = spacy_experimental.edit_tree_lemmatizer.edit_tree_lemmatizer:make_edit_tree_lemmatizer
     experimental_char_tagger_tokenizer = spacy_experimental.char_tokenizer.char_tagger_tokenizer:make_char_tagger_tokenizer
     experimental_char_ner_tokenizer = spacy_experimental.char_tokenizer.char_ner_tokenizer:make_char_ner_tokenizer
