--- spacy_experimental/biaffine_parser/pyproject.toml
@@ -1,8 +1,0 @@
-[build-system]
-requires = [
-  "setuptools",
-  "cython>=0.25,<3.0",
-  "wheel",
-  "numpy>=1.15.0",
-  "spacy>=3.2.0,<3.3.0"
-]
