--- spacy_experimental/biaffine_parser/setup.cfg
@@ -1,37 +1,0 @@
-[metadata]
-version = 0.0.1
-author = Explosion
-author_email = contact@explosion.ai
-license = MIT
-long_description = file: README.md
-long_description_content_type = text/markdown
-
-[options]
-zip_safe = false
-python_requires = >=3.6
-setup_requires =
-    cython>=0.25,<3.0
-    numpy
-    spacy>=3.2.0,<3.3.0
-install_requires =
-    numpy
-    spacy>=3.2.0,<3.3.0
-
-[options.entry_points]
-spacy_factories =
-    arc_predicter = spacy_biaffine_parser.arc_predicter:make_arc_predicter
-    arc_labeler = spacy_biaffine_parser.arc_labeler:make_arc_labeler
-    Bilinear.v1 = spacy_biaffine_parser.pairwise_bilinear:build_pairwise_bilinear
-    PairwiseBilinear.v1 = spacy_biaffine_parser.bilinear:build_bilinear
-
-[bdist_wheel]
-universal = false
-
-[sdist]
-formats = gztar
-
-[mypy]
-ignore_missing_imports = True
-no_implicit_optional = True
-plugins = pydantic.mypy, thinc.mypy
-allow_redefinition = True
